+++ Tests/Utilities/LSLFramework.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b7f9237a0219ab54387912dd2cfb35c7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
