+++ Tests/Utilities/LSLFramework/LSLRawStreamReader.cs.meta
